requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai
gunicorn==21.2.0
gspread
google-auth
